Introduce a dev dependency on elm-format
What?
=====

This introduces a development dependency on `elm-format`; with this, all
subsequent changes to any Elm files in the repository should use the
locked version of `elm-format` to format the files to ensure consistency
across all developers' Elm changes.

This also updates all files with changes to this version of the
formatter.