Fix #65 add highlights for fuzzy matches
For users of MS Excel we can highlight differences between
canonical forms used used for fuzzy matching. We use the
damerau-levenshtein gem >= v1.3.0 for this purpose. For now
the gem calculates difference using ruby only. If we see
significant perforance problem, we will convert time-consuming
parts of the algorithm to C language.